переделал TrendService

+++ src/ru/ex4/apibt/service/TrendServiceTest.java
@@ -1,5 +1,6 @@
 package ru.ex4.apibt.service;
 
+import junit.framework.Assert;
 import org.junit.Test;
 import ru.ex4.apibt.dto.TrendType;
 
@@ -20,44 +21,44 @@ public class TrendServiceTest {
     }
 
     private void testEmpty() throws IOException {
-        ArrayList<Float> lastTradeList = new ArrayList<>();
+        ArrayList<Integer> lastTradeList = new ArrayList<>();
         TrendType trendType = TrendService.extractType(lastTradeList);
         assertNull(trendType);
     }
 
     private void testOnUpward() throws IOException {
-        ArrayList<Float> lastTradeList1 = new ArrayList<>();
-        lastTradeList1.add(10F);
-        lastTradeList1.add(12F);
-        lastTradeList1.add(18F);
+        ArrayList<Integer> lastTradeList1 = new ArrayList<>();
+        lastTradeList1.add(10);
+        lastTradeList1.add(12);
+        lastTradeList1.add(18);
         TrendType trendType1 = TrendService.extractType(lastTradeList1);
         assertTrue(trendType1 == TrendType.upward);
     }
 
     private void testDownward() throws IOException {
-        ArrayList<Float> lastTradeList2 = new ArrayList<>();
-        lastTradeList2.add(19F);
-        lastTradeList2.add(12F);
-        lastTradeList2.add(8F);
+        ArrayList<Integer> lastTradeList2 = new ArrayList<>();
+        lastTradeList2.add(19);
+        lastTradeList2.add(12);
+        lastTradeList2.add(8);
         TrendType trendType2 = TrendService.extractType(lastTradeList2);
         assertTrue(trendType2 == TrendType.downward);
     }
 
     private void testFlat() throws IOException {
-        ArrayList<Float> lastTradeList3 = new ArrayList<>();
-        lastTradeList3.add(25F);
-        lastTradeList3.add(18F);
-        lastTradeList3.add(20F);
+        ArrayList<Integer> lastTradeList3 = new ArrayList<>();
+        lastTradeList3.add(25);
+        lastTradeList3.add(18);
+        lastTradeList3.add(20);
 
         TrendType trendType3 = TrendService.extractType(lastTradeList3);
         assertTrue(trendType3 == TrendType.flat);
     }
 
     private void testFlat2() throws IOException {
-        ArrayList<Float> lastTradeList3 = new ArrayList<>();
-        lastTradeList3.add(20F);
-        lastTradeList3.add(20F);
-        lastTradeList3.add(20F);
+        ArrayList<Integer> lastTradeList3 = new ArrayList<>();
+        lastTradeList3.add(20);
+        lastTradeList3.add(20);
+        lastTradeList3.add(20);
 
         TrendType trendType3 = TrendService.extractType(lastTradeList3);
         assertTrue(trendType3 == TrendType.flat);
